requirements.txt
fastapi
uvicorn[standard]
httpx
PyYAML
slowapi
scikit-learn
aiosqlite
jinja2
python-multipart

